+++ src/libs/libs/round_trip_time.h
@@ -3,7 +3,7 @@
 
 typedef struct RoundTripTime RoundTripTime;
 
-RoundTripTime *round_trip_time_construct(int src, int dest, double rtt);
+RoundTripTime *round_trip_time_construct(int src, int dest, long double rtt);
 
 void round_trip_time_destruct(RoundTripTime *round_trip_time);
 
@@ -11,7 +11,7 @@ int round_trip_time_get_src(RoundTripTime *round_trip_time);
 
 int round_trip_time_get_dest(RoundTripTime *round_trip_time);
 
-double round_trip_time_get_rtt(RoundTripTime *round_trip_time);
+long double round_trip_time_get_rtt(RoundTripTime *round_trip_time);
 
 int round_trip_compare(const void *a, const void *b);
